test: move producer test functions to separate test file
